refactor: move spinner to tui components
- Moved spinner from format package to tui/components/spinner
- Updated imports and references in root.go
- Added basic test for spinner component
